Move track state creator etc. out of the CKF.
Merge retrieval of source link ranges, measurement selection and
track state creation into one unit which the CKF interacts with
via a single delegate. The delegates for measurement selection
track state creation and calibration are removed from the CKF
options/extensions.

The original building blocks (SourceLink accessor, measurement selector,
and track state creator) can still be used, however require to setup an
additional helper object which combines these independent components. The
algorithmic code is unchanged.
